Add beat-triggered drops mode with explicit, time-based afterglow
- Move the per-frame brightness decay out of LedString::paint_string into
  Canvas::fade(dt, tau). Effects that want trails call it explicitly; the
  existing modes redraw every pixel and are unchanged.
- Add DisplayMode::Mic_Drops: on a detected beat a bright head runs down
  every tentacle, the ring and noodles flash and decay. Only the drop head
  is drawn, the tail comes from the fade. A lone drop keeps the jelly alive
  in silence.
- Run the LED channel test on the wall clock instead of a frame counter.
- Make LedString and Microphone non-copyable; they own buffers, a DMA
  channel and PIO state machines.
- Clamp PwmLight::set_level to 0..1.
- Do the 24-bit sign extension of microphone samples on an unsigned copy.

+++ firmware_cpp/jell_audio.cpp
@@ -117,7 +117,8 @@ AudioFrame Microphone::capture()
 
         // Shift left to force the 24th bit into the 32nd bit slot, 
         // then arithmetic shift right back down to sign-extend automatically.
-        samples[i] = (samples[i] << 8) >> 8;
+        // The left shift happens on the unsigned copy, which is well defined for any value.
+        samples[i] = (int32_t)((uint32_t)samples[i] << 8) >> 8;
     
         sum += samples[i];
     }
